Separated mm.h public from mm_private.h
